+++ cpp/tests/two_process_sub.cpp
@@ -34,7 +34,7 @@ int main(int argc, char** argv) {
         return 2;
     }
 
-    uint16_t topic_id = static_cast<uint16_t>(std::atoi(argv[argi++]));
+    uint32_t topic_id = static_cast<uint32_t>(std::atoi(argv[argi++]));
     int expected_count = std::atoi(argv[argi++]);
     int timeout_sec = std::atoi(argv[argi++]);
